build: pam_p11: unsigned comparison

+++ src/pam_p11.c
@@ -553,7 +553,7 @@ static int randomize(pam_handle_t *pamh, unsigned char *r, unsigned int r_len)
 {
 	int ok = 0;
 	int fd = open("/dev/urandom", O_RDONLY);
-	if (0 <= fd && read(fd, r, r_len) == r_len) {
+	if (0 <= fd && read(fd, r, r_len) == (ssize_t)r_len) {
 		ok = 1;
 	} else {
 		pam_syslog(pamh, LOG_CRIT, "Error reading from /dev/urandom: %s",
